+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Improved error reporting during actor document parsing.
 - Optimized post queries.
 - Improved documentation of `list-filter-rules` command.
+- Retry incoming activities only on fetcher errors.
 
 ## [3.7.1] - 2024-10-17